# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Live Site AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      tags: [admin]
      operationId: getAdminUsers
      summary: List all users (admin)
      parameters:
        - { name: page, in: query, required: false, schema: { type: integer } }
        - { name: pageSize, in: query, required: false, schema: { type: integer } }
        - { name: search, in: query, required: false, schema: { type: string } }
        - { name: status, in: query, required: false, schema: { type: string, enum: [pending_approval, active, suspended] } }
      responses:
        '200':
          description: Paginated list of users
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminUsersList' }
        '403': { description: Forbidden }
    post:
      tags: [admin]
      operationId: manualCreateUser
      summary: Manually create or upsert a user by email (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ManualUserUpsertBody' }
      responses:
        '200':
          description: Existing user updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ManualUserUpsertResponse' }
        '201':
          description: New user created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ManualUserUpsertResponse' }
  /admin/users/{id}/approve:
    post:
      tags: [admin]
      operationId: approveUser
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: User approved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessEnvelope' }
  /admin/users/{id}/reject:
    post:
      tags: [admin]
      operationId: rejectUser
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: User rejected (suspended)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessEnvelope' }
  /admin/users/{id}/set-tier:
    post:
      tags: [admin]
      operationId: setUserTier
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetTierBody' }
      responses:
        '200':
          description: Tier updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessEnvelope' }
  /admin/users/{id}/set-role:
    post:
      tags: [admin]
      operationId: setUserRole
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetRoleBody' }
      responses:
        '200':
          description: Role updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessEnvelope' }
  /admin/invites:
    get:
      tags: [admin]
      operationId: listInvites
      parameters:
        - { name: page, in: query, required: false, schema: { type: integer } }
        - { name: pageSize, in: query, required: false, schema: { type: integer } }
      responses:
        '200':
          description: Paginated invites
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminInvitesList' }
    post:
      tags: [admin]
      operationId: createInvite
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateInviteBody' }
      responses:
        '201':
          description: Invite created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreateInviteResponse' }
  /admin/invites/{id}/revoke:
    post:
      tags: [admin]
      operationId: revokeInvite
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Invite revoked
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessEnvelope' }
  /admin/demos:
    get:
      tags: [admin]
      operationId: getAdminDemos
      summary: List all demos across all users (admin)
      parameters:
        - { name: page, in: query, required: false, schema: { type: integer } }
        - { name: pageSize, in: query, required: false, schema: { type: integer } }
        - { name: search, in: query, required: false, schema: { type: string } }
      responses:
        '200':
          description: Paginated list of demos
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminDemosList' }
        '403': { description: Forbidden }
  /admin/users/{id}/suspend:
    post:
      tags: [admin]
      operationId: suspendUser
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: User suspended
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessEnvelope' }
  /admin/users/{id}/reactivate:
    post:
      tags: [admin]
      operationId: reactivateUser
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: User reactivated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessEnvelope' }
  /admin/audit-log:
    get:
      tags: [admin]
      operationId: getAdminAuditLog
      parameters:
        - { name: page, in: query, required: false, schema: { type: integer } }
        - { name: pageSize, in: query, required: false, schema: { type: integer } }
      responses:
        '200':
          description: Paginated audit log
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminAuditLogList' }
  /admin/impersonate/{userId}:
    post:
      tags: [admin]
      operationId: startImpersonation
      parameters:
        - { name: userId, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Impersonation started
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessEnvelope' }
...
components:
...
  schemas:
...
    AdminUserRow:
      type: object
      required: [id, email, firstName, lastName, role, status, tier, createdAt, demoCount, demosCreated, enrichmentsToday, totalEnrichments, lastActivity, lastLoginAt]
      properties:
        id: { type: string }
        email: { type: [string, 'null'] }
        firstName: { type: [string, 'null'] }
        lastName: { type: [string, 'null'] }
        role: { type: string }
        status: { type: string }
        tier: { type: string, enum: [free, pro] }
        createdAt: { type: string, format: date-time }
        demoCount: { type: integer }
        demosCreated: { type: integer }
        enrichmentsToday: { type: integer }
        totalEnrichments: { type: integer }
        lastActivity:
          oneOf:
            - { type: string, format: date-time }
            - { type: 'null' }
        lastLoginAt:
          oneOf:
            - { type: string, format: date-time }
            - { type: 'null' }
    AdminUsersList:
      type: object
      required: [page, pageSize, total, items]
      properties:
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
        items:
          type: array
          items: { $ref: '#/components/schemas/AdminUserRow' }
    AdminDemoRow:
      type: object
      required: [id, companyName, slug, status, viewCount, callClickCount, calendarClickCount, createdAt, userId, ownerEmail]
      properties:
        id: { type: string }
        companyName: { type: string }
        slug: { type: string }
        status: { type: string }
        viewCount: { type: integer }
        callClickCount: { type: integer }
        calendarClickCount: { type: integer }
        createdAt: { type: string, format: date-time }
        userId: { type: string }
        ownerEmail: { type: [string, 'null'] }
    AdminDemosList:
      type: object
      required: [page, pageSize, total, items]
      properties:
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
        items:
          type: array
          items: { $ref: '#/components/schemas/AdminDemoRow' }
    AdminAuditLogRow:
      type: object
      required: [id, actorId, action, createdAt]
      properties:
        id: { type: string }
        actorId: { type: string }
        actorEmail: { type: [string, 'null'] }
        action: { type: string }
        targetType: { type: [string, 'null'] }
        targetId: { type: [string, 'null'] }
        details: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
    AdminAuditLogList:
      type: object
      required: [page, pageSize, total, items]
      properties:
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
        items:
          type: array
          items: { $ref: '#/components/schemas/AdminAuditLogRow' }
...
    AdminInviteRow:
      type: object
      required: [id, token, tier, createdAt, createdBy]
      properties:
        id: { type: string }
        token: { type: string }
        tier: { type: string, enum: [free, pro] }
        invitedEmail: { type: [string, 'null'] }
        note: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
        expiresAt:
          oneOf:
            - { type: string, format: date-time }
            - { type: 'null' }
        consumedAt:
          oneOf:
            - { type: string, format: date-time }
            - { type: 'null' }
        consumedByUserId: { type: [string, 'null'] }
        revokedAt:
          oneOf:
            - { type: string, format: date-time }
            - { type: 'null' }
        createdBy: { type: string }
    AdminInvitesList:
      type: object
      required: [page, pageSize, total, items]
      properties:
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
        items:
          type: array
          items: { $ref: '#/components/schemas/AdminInviteRow' }
    CreateInviteBody:
      type: object
      required: [tier]
      properties:
        tier: { type: string, enum: [free, pro] }
        invitedEmail: { type: [string, 'null'] }
        expiresInDays: { type: [integer, 'null'] }
        note: { type: [string, 'null'] }
    CreateInviteResponse:
      type: object
      required: [invite]
      properties:
        invite: { $ref: '#/components/schemas/AdminInviteRow' }
    ManualUserUpsertBody:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
        tier: { type: string, enum: [free, pro] }
        role: { type: string, enum: [user, admin] }
        firstName: { type: [string, 'null'] }
        lastName: { type: [string, 'null'] }
    ManualUserUpsertResponse:
      type: object
      required: [user, created]
      properties:
        user: { $ref: '#/components/schemas/AdminUserRow' }
        created: { type: boolean }
    SetTierBody:
      type: object
      required: [tier]
      properties:
        tier: { type: string, enum: [free, pro] }
    SetRoleBody:
      type: object
      required: [role]
      properties:
        role: { type: string, enum: [user, admin] }
    SuccessEnvelope:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...